# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for all fests",
        "title": "FestAPI v2",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/user/dauth/callback": {
            "get": {
                "description": "Callback url for DAuth, returns JWT token if successful",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Authenticate user with DAuth",
                "operationId": "DAuthUserLogin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAuth code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/details": {
            "get": {
                "security": [
                    {
                        "middleware.UserAuth": []
                    }
                ],
                "description": "profile information to be displayed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Profile information.",
                "operationId": "ProfileDetails",
                "responses": {
                    "400": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "Authenticates a user using email and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Authenticate and log in a user.",
                "operationId": "AuthUserLogin",
                "parameters": [
                    {
                        "description": "User authentication request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/qr": {
            "get": {
                "security": [
                    {
                        "middleware.UserAuth": []
                    }
                ],
                "description": "QR for the profile page.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "QR Generation.",
                "operationId": "ProfileQR",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Register a new user with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a new user.",
                "operationId": "AuthUserRegister",
                "parameters": [
                    {
                        "description": "User registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/update": {
            "patch": {
                "security": [
                    {
                        "middleware.UserAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update user information with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user information.",
                "operationId": "AuthUserUpdate",
                "parameters": [
                    {
                        "description": "User update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AuthUserLoginRequest": {
            "type": "object",
            "required": [
                "user_email",
                "user_password"
            ],
            "properties": {
                "user_email": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                }
            }
        },
        "dto.AuthUserRegisterRequest": {
            "type": "object",
            "properties": {
                "user_address": {
                    "type": "string"
                },
                "user_city": {
                    "type": "string"
                },
                "user_college": {
                    "type": "string"
                },
                "user_country": {
                    "type": "string"
                },
                "user_degree": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                },
                "user_fullname": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_nationality": {
                    "type": "string"
                },
                "user_othercollege": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                },
                "user_phone": {
                    "type": "string"
                },
                "user_pincode": {
                    "type": "string"
                },
                "user_recaptcha_code": {
                    "type": "string"
                },
                "user_referral_code": {
                    "type": "string"
                },
                "user_sex": {
                    "type": "string"
                },
                "user_sponsor": {
                    "type": "string"
                },
                "user_state": {
                    "type": "string"
                },
                "user_voucher_name": {
                    "type": "string"
                },
                "user_year": {
                    "type": "string"
                }
            }
        },
        "dto.AuthUserUpdateRequest": {
            "type": "object",
            "properties": {
                "user_address": {
                    "type": "string"
                },
                "user_city": {
                    "type": "string"
                },
                "user_college": {
                    "type": "string"
                },
                "user_country": {
                    "type": "string"
                },
                "user_degree": {
                    "type": "string"
                },
                "user_nationality": {
                    "type": "string"
                },
                "user_othercollege": {
                    "type": "string"
                },
                "user_phone": {
                    "type": "string"
                },
                "user_pincode": {
                    "type": "string"
                },
                "user_referral_code": {
                    "type": "string"
                },
                "user_sex": {
                    "type": "string"
                },
                "user_sponsor": {
                    "type": "string"
                },
                "user_state": {
                    "type": "string"
                },
                "user_voucher_name": {
                    "type": "string"
                },
                "user_year": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}